refactor(step4):UrlMapping하는 기능과 User 기능을 분리하기 위해 UrlMapper 객체 구현
