+++ src/main/java/controller/UrlMapper.java
@@ -1,0 +1,29 @@
+package controller;
+
+import request.HttpRequest;
+
+public class UrlMapper {
+    private UserController userController;
+
+    public UrlMapper(UserController userController) {
+        this.userController = userController;
+    }
+
+    /**
+     * HTTP Request를 받아 요청 URL을 분석해 비즈니스 로직을 수행하고 뷰를 String으로 반환합니다.
+     * @param httpRequest
+     * @return view
+     */
+    public String requestMapping(HttpRequest httpRequest) {
+        if (httpRequest.getUrl().equals("/")) {
+            httpRequest.setUrl("/index.html");
+            return httpRequest.getUrl();
+        }
+
+        if (httpRequest.getUrl().equals("/user/create")) {
+            return userController.userJoin(httpRequest);
+        }
+
+        return httpRequest.getUrl();
+    }
+}
